back-end/supplier-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
pydantic[email]==2.9.2
email-validator==2.2.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
